refactor: context에서 hook 분리에 따른 의존성 수정
