+++ client/src/App.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import GlobalStyle from './styles/GlobalStyle';
 import { BrowserRouter } from 'react-router-dom';
 import Router from '@/Router';
-import { UserProvider } from '@/contexts/User';
+import { UserProvider } from '@/contexts/userContext';
 
 function App() {
   return (
